fix manual sankaku tagging

+++ classes/api.py
@@ -63,7 +63,7 @@ class API:
         except Exception as e:
             print(f'Could not process your query: {e}.')
 
-    def get_post(self, post_id, local_temp_path=None):
+    def get_post(self, post_id, local_temp_path=None, sankaku_url=None):
         """
         Returns a boilerplate post object with post_id, image_url and version.
 
@@ -78,6 +78,7 @@ class API:
         """
 
         try:
+            blacklist_extensions = ['mp4', 'webm', 'mkv']
             query_url   = self.booru_api_url + '/post/' + post_id
             response    = requests.get(query_url, headers=self.headers)
 
@@ -93,22 +94,25 @@ class API:
 
             # Download image and add it to the post object
             # ToDo: Don't do that if the booru is accessible over the internet
-            filename = content_url.split('/')[-1]
-            local_file_path = urllib.request.urlretrieve(image_url, local_temp_path + filename)[0]
+            if not any(extension in content_url for extension in blacklist_extensions):
+                filename = content_url.split('/')[-1]
+                local_file_path = urllib.request.urlretrieve(image_url, local_temp_path + filename)[0]
 
-            # Resize image if it's too big. IQDB limit is 8192KB or 7500x7500px.
-            # Resize images bigger than 3MB to reduce stress on iqdb.
-            image_size = os.path.getsize(local_file_path)
+                # Resize image if it's too big. IQDB limit is 8192KB or 7500x7500px.
+                # Resize images bigger than 3MB to reduce stress on iqdb.
+                image_size = os.path.getsize(local_file_path)
 
-            if image_size > 3000000:
-                resize_image(local_file_path)
+                if image_size > 3000000:
+                    resize_image(local_file_path)
 
-            with open(local_file_path, 'rb') as f:
-                image = f.read()
+                with open(local_file_path, 'rb') as f:
+                    image = f.read()
 
-            # Remove temporary image
-            if os.path.exists(local_file_path):
-                os.remove(local_file_path)
+                # Remove temporary image
+                if os.path.exists(local_file_path):
+                    os.remove(local_file_path)
+            else:
+                image = None
 
             post = Post(md5sum, post_id, image_url, image, version, tag_list)
 
@@ -131,7 +135,8 @@ class API:
         meta_data = json.dumps({"version": post.version, "tags": post.tags, "source": post.source, "safety": post.rating})
 
         try:
-            requests.put(query_url, headers=self.headers, data=meta_data)
+            response = requests.put(query_url, headers=self.headers, data=meta_data)
+            print(response.json())
         except Exception as e:
             print(f'Could not upload your post: {e}')
